logout and delete confirmation

+++ src/app/ProjectComponents/CommonComponents/home/home.component.ts
@@ -1,6 +1,8 @@
 import { OnInit } from '@angular/core';
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { LogOffComponent } from '../../log-off/log-off.component';
 
 @Component({
   selector: 'app-home',
@@ -8,7 +10,7 @@ import { Router } from '@angular/router';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit{
-  constructor(private router:Router){}
+  constructor(private router:Router,public dialog:MatDialog){}
   ngOnInit(): void {
     if(sessionStorage.getItem('role')!=='ADMIN' || sessionStorage.getItem('role')===null){
       this.router.navigate([""]);
@@ -16,8 +18,14 @@ export class HomeComponent implements OnInit{
   }
 
   logOut(){
-    alert("Logging off");
-    sessionStorage.clear();
-    this.router.navigate([""]);
+    // alert("Logging off");
+    const dialogRef= this.dialog.open(LogOffComponent)
+    dialogRef.afterClosed().subscribe((res)=>{
+      if(res=='yes')
+      {
+      sessionStorage.clear();
+      this.router.navigate([""]);
+      }
+    })
   }
 }
